+++ src/controllers/user/verify-email.ts
@@ -8,6 +8,7 @@ import { isBrowserTrustedForUser } from "../../managers/browser-authentication";
 import {
   addAuthenticationMethodReferenceInSession,
   getUserFromAuthenticatedSession,
+  isWithinAuthenticatedSession,
 } from "../../managers/session/authenticated";
 import {
   sendEmailAddressVerificationEmail,
@@ -117,6 +118,30 @@ export const postSendEmailVerificationController = async (
       );
     }
 
+    next(error);
+  }
+};
+
+export const getVerificationCodeController = async (
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) => {
+  try {
+    let email: string | undefined;
+    let user: User | undefined;
+
+    if (isWithinAuthenticatedSession(req.session)) {
+      user = getUserFromAuthenticatedSession(req);
+      email = user.email;
+    }
+
+    return res.render("user/verify-email-help", {
+      pageTitle: "Renvoyer un code",
+      email,
+      csrfToken: email && csrfToken(req),
+    });
+  } catch (error) {
     next(error);
   }
 };
