+++ src/views/user/verify-email.ejs
@@ -47,7 +47,7 @@
                 <div>
                     <a
                         class="fr-link fr-icon-question-line fr-link--icon-left"
-                        href="/users/verification-code"
+                        href="/users/verify-email-help"
                     >
                         J'ai attendu 15 minutes et je ne reçois pas de code de vérification
                     </a>
